requirements.txt
python-telegram-bot
google-genai
httpx
certifi
certifi
dnspython